Rename & move downloadQueue
