+++ mobile/lib/utils/file_download_util.dart
@@ -277,37 +277,4 @@ Future<void> _saveLivePhotoOnDroid(
   );
   await IgnoredFilesService.instance.cacheAndInsert([ignoreVideoFile]);
 }
-
-class DownloadQueue {
-  final int maxConcurrent;
-  final Queue<Future<void> Function()> _queue = Queue();
-  int _runningTasks = 0;
-
-  DownloadQueue({this.maxConcurrent = 5});
-
-  Future<void> add(Future<void> Function() task) async {
-    final completer = Completer<void>();
-    _queue.add(() async {
-      try {
-        await task();
-        completer.complete();
-      } catch (e) {
-        completer.completeError(e);
-      } finally {
-        _runningTasks--;
-        _processQueue();
-      }
-      return completer.future;
-    });
-    _processQueue();
-    return completer.future;
-  }
-
-  void _processQueue() {
-    while (_runningTasks < maxConcurrent && _queue.isNotEmpty) {
-      final task = _queue.removeFirst();
-      _runningTasks++;
-      task();
-    }
-  }
 }
